code/requirements.txt
# Spanish Corpus Scraping Framework - Production Dependencies
# Generated for enterprise deployment and reproducible environments

# Core HTTP and Web Scraping
requests==2.31.0
urllib3==2.0.7

# HTML Parsing and Content Extraction
beautifulsoup4==4.12.3
lxml==5.1.0
trafilatura==1.9.0
html-sanitizer==1.9.3

# Configuration Management
PyYAML==6.0.1

# Politeness and Resilience
protego==0.4.0
tenacity==8.2.3
fake-useragent==1.5.1

# Natural Language Processing and Validation
spacy==3.7.4
fasttext-wheel==0.9.2

# RSS and Feed Processing
feedparser==6.0.11

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Optional: Headless Browser Support
playwright==1.44.0

python-dotenv

# Note: Additional setup required after pip install:
# 1. Download spaCy Spanish model:
#    python -m spacy download es_core_news_sm
# 
# 2. Install Playwright browsers (if using JavaScript-heavy sites):
#    playwright install
#
# 3. Download fasttext language detection model:
#    wget https://dl.fbaipublicfiles.com/fasttext/supervised-models/lid.176.bin
